% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spData.R
\docType{data}
\name{spData}
\alias{spData}
\title{Mid-infrared Spectra of Thirteen Types of Starch}
\format{A data frame with 7469 observations of 14 variables}
\usage{
data(spData)
}
\description{
A dataset containing mid-infrared spectra of thirteen types of starch
}
\details{
\itemize{
  \item Wavenumber. Wavenumber of mid-infrared spectra.
  \item A-M. Absorbance of an mid-infrared spectra of thirteen types of starch (A.U.).
}
}
\keyword{datasets}

